torbox-strm/requirements.txt
apscheduler
tinydb
httpx
python-dotenv
parse-torrent-title
fuse-python; sys_platform != "win32" and sys_platform != "darwin"
git+https://github.com/libfuse/python-fuse; sys_platform == "darwin"
